refactor(imagewithplaceholder): make the image-loading state unique
Previously, we have two states, one for whether the image is loaded and the other for the loading
error. To make it explicit that these two states are mutually dependent (i.e. we cannot have both
loaded and error), we now manage the loading state as one variable that takes either of the three
values, loading, loaded, and error. For handling the case where the src attribute is missing, we use
prop-types: the missing src should be reported to the developer, not to the user.