fix(imagewithplaceholder): trigger the onload event when the image is cached
Previously, when the image is cached, the onload event won't be fired. As a result, the "Loading.."
message hides the image even though the latter is loaded. We use the imageElement.complete property
to check if the image is cached, and if so, we set the image-loading status to be "loaded".